+++ frontend/app/api/banks/stgeorge.ts
@@ -1,0 +1,4 @@
+export type StGeorgeBankAccountDetails = {
+  bsbNumber: string;
+  accountNumber: string;
+};
