+++ frontend/app/api/banks/westpac.ts
@@ -1,0 +1,4 @@
+export type WestpacBankAccountDetails = {
+  bsbNumber: string;
+  accountNumber: string;
+};
