+++ frontend/app/api/environment.ts
@@ -1,0 +1,1 @@
+export type Environment = "development" | "test" | "production";
