+++ frontend/app/components/accounts/SyncDirectionIcon.tsx
@@ -1,0 +1,5 @@
+import { ArrowRightIcon } from "@heroicons/react/24/outline";
+
+export function SyncDirectionIcon() {
+  return <ArrowRightIcon className="h-4 w-4 text-gray-500" />;
+}
