+++ frontend/app/components/accounts/YnabAccountSummary.tsx
@@ -1,0 +1,21 @@
+import { YnabAccountDetail } from "~/api/api";
+import { YnabAccountSummary as YnabAccountSummaryApi } from "~/api/accountSummary";
+import { YnabIcon } from "../ynab/YnabIcon";
+
+export function YnabAccountSummary({
+  account: ynabAccount,
+}: {
+  account: YnabAccountDetail | YnabAccountSummaryApi;
+}) {
+  return (
+    <div className="flex items-center gap-4">
+      <YnabIcon />
+      <div className="flex flex-col">
+        <div className="text-gray-700 dark:text-gray-300">
+          {ynabAccount.accountName}
+        </div>
+        <div className="text-sm text-gray-500">{ynabAccount.budgetName}</div>
+      </div>
+    </div>
+  );
+}
