+++ frontend/app/routes/accounts/index.tsx
@@ -1,14 +1,11 @@
 import { MagnifyingGlassIcon } from "@heroicons/react/24/outline";
 import { json } from "@remix-run/node";
 import { NavLink, useLoaderData } from "@remix-run/react";
-import classnames from "classnames";
-import type { AccountDetail } from "~/api/api";
-import { getSyncDetails } from "~/api/api";
-import {
-  BankAccountSummary,
-  SyncDirectionIcon,
-  YnabAccountSummary,
-} from "~/components/accounts/AccountSummary";
+import type { AccountSummary } from "~/api/accountSummary";
+import { getAccountSummaries } from "~/api/api";
+import { SyncDirectionIcon } from "~/components/accounts/SyncDirectionIcon";
+import { YnabAccountSummary } from "~/components/accounts/YnabAccountSummary";
+import { BankAccountSummary } from "~/components/accounts/BankAccountSummary";
 import { useRefreshOnInterval } from "~/components/hooks/useRefreshOnInterval";
 import { Paper } from "~/components/layout/Paper";
 import { Heading } from "~/components/primitive/Heading";
@@ -17,11 +14,11 @@ import { SyncStatusWithLastSyncTime } from "~/components/sync/SyncStatus";
 import { ContentHeader } from "../../components/layout/ContentHeader";
 
 export async function loader() {
-  return json(await getSyncDetails());
+  return json(await getAccountSummaries());
 }
 
 export default function Accounts() {
-  const data = useLoaderData<Array<AccountDetail>>();
+  const accounts = useLoaderData<Array<AccountSummary>>();
   useRefreshOnInterval({ enabled: true, interval: 5000 });
 
   return (
@@ -29,43 +26,35 @@ export default function Accounts() {
       <ContentHeader>
         <div className="flex w-full items-center gap-4">
           <Heading title="Accounts" />
-          <div className="relative ml-auto hidden text-gray-400">
-            <div className="pointer-events-none absolute flex h-10 w-10 items-center justify-center">
-              <MagnifyingGlassIcon className="h-5 w-5 stroke-gray-600" />
-            </div>
-            <input
-              type="text"
-              placeholder="Search accounts"
-              className="h-10 rounded-full border-0 border-gray-600 bg-gray-800 px-4 pl-10 text-sm text-gray-400 placeholder:text-gray-600 hover:bg-gray-700 hover:placeholder:text-gray-500 focus:border-gray-600 focus:bg-gray-700 focus:ring-0 focus:placeholder:text-gray-400"
-            />
-          </div>
         </div>
       </ContentHeader>
       <div className="container mx-auto gap-2">
         <div className="grid grid-cols-1 gap-4 lg:grid-cols-2 xl:grid-cols-3">
-          {data.map((d) => {
+          {accounts.map((account) => {
             return (
-              <Paper key={d.id}>
+              <Paper key={account.id}>
                 <NavLink
-                  key={d.id}
-                  to={`/accounts/${d.id}`}
+                  key={account.id}
+                  to={`/accounts/${account.id}`}
                   className="text flex flex-col gap-4 rounded-lg"
                 >
                   <div className="flex items-center gap-4">
-                    <BankAccountSummary account={d.bank} />
+                    <BankAccountSummary account={account.bankAccount} />
                     <SyncDirectionIcon />
-                    <YnabAccountSummary account={d.ynab} />
+                    <YnabAccountSummary account={account.ynabAccount} />
                   </div>
                   <div className="flex items-center gap-2">
                     <SyncStatusWithLastSyncTime
-                      status={d.status}
+                      status={account.status}
                       lastSyncTime={
-                        d.lastSyncTime ? new Date(d.lastSyncTime) : undefined
+                        account.lastSyncTime
+                          ? new Date(account.lastSyncTime)
+                          : undefined
                       }
                     />
                   </div>
                   <div className="-ml-3 flex items-center">
-                    <SyncNowButton accountId={d.id} />
+                    <SyncNowButton accountId={account.id} />
                   </div>
                 </NavLink>
               </Paper>
