+++ frontend/package.json
@@ -37,6 +37,7 @@
     "@remix-run/dev": "^1.7.2",
     "@remix-run/eslint-config": "^1.7.2",
     "@tailwindcss/forms": "^0.5.3",
+    "@types/lodash": "4.14.189",
     "@types/react": "^18.0.15",
     "@types/react-dom": "^18.0.6",
     "eslint": "^8.23.1",
